tests: Move Eloquent relationship tests into Repository
Also, remove transaction around deletion test; the transaction is
implemented in the BaseRepository's delete() method now.
